distributed fixes
