+++ cellarium/ml/utilities/data.py
@@ -8,7 +8,6 @@ Data utilities
 This module contains helper functions for data loading and processing.
 """
 
-import warnings
 from collections.abc import Callable
 from dataclasses import dataclass
 from typing import Any
@@ -17,10 +16,8 @@ import numpy as np
 import pandas as pd
 import scipy
 import torch
-import torch.distributed as dist
 from anndata import AnnData
 from anndata.experimental import AnnCollection
-from torch.utils.data import get_worker_info as _get_worker_info
 
 
 @dataclass
@@ -76,54 +73,6 @@ class AnnDataField:
         if self.attr == "obs":
             result = self.key
         return result
-
-
-def get_rank_and_num_replicas() -> tuple[int, int]:
-    """
-    This helper function returns the rank of the current process and
-    the number of processes in the default process group. If distributed
-    package is not available or default process group has not been initialized
-    then it returns ``rank=0`` and ``num_replicas=1``.
-
-    Returns:
-        Tuple of ``rank`` and ``num_replicas``.
-    """
-    if not dist.is_available():
-        num_replicas = 1
-        rank = 0
-    else:
-        try:
-            num_replicas = dist.get_world_size()
-            rank = dist.get_rank()
-        except (ValueError, RuntimeError):  # RuntimeError was changed to ValueError in PyTorch 2.2
-            warnings.warn(
-                "Distributed package is available but the default process group has not been initialized. "
-                "Falling back to ``rank=0`` and ``num_replicas=1``.",
-                UserWarning,
-            )
-            num_replicas = 1
-            rank = 0
-    if rank >= num_replicas or rank < 0:
-        raise ValueError(f"Invalid rank {rank}, rank should be in the interval [0, {num_replicas-1}]")
-    return rank, num_replicas
-
-
-def get_worker_info() -> tuple[int, int]:
-    """
-    This helper function returns ``worker_id`` and ``num_workers``. If it is running
-    in the main process then it returns ``worker_id=0`` and ``num_workers=1``.
-
-    Returns:
-        Tuple of ``worker_id`` and ``num_workers``.
-    """
-    worker_info = _get_worker_info()
-    if worker_info is None:
-        worker_id = 0
-        num_workers = 1
-    else:
-        worker_id = worker_info.id
-        num_workers = worker_info.num_workers
-    return worker_id, num_workers
 
 
 def collate_fn(batch: list[dict[str, np.ndarray]]) -> dict[str, np.ndarray | torch.Tensor]:
