--- docs/source/distributed.rst
@@ -1,7 +1,0 @@
-Distributed
-===========
-
-.. automodule:: cellarium.ml.distributed
-   :members:
-   :special-members:
-   :show-inheritance:
