+++ docs/source/utilities.rst
@@ -7,6 +7,11 @@ Utilities
    :show-inheritance:
    :exclude-members: __weakref__
 
+.. automodule:: cellarium.ml.utilities.distributed
+   :members:
+   :special-members:
+   :show-inheritance:
+
 .. automodule:: cellarium.ml.utilities.testing
    :members:
    :special-members:
