+++ pyproject.toml
@@ -27,7 +27,7 @@ dependencies = [
   "crick>=0.0.4",
   "google-cloud-storage",
   "jsonargparse[signatures]==4.27.7",
-  "lightning>=2.2.0",
+  "lightning>=2.2.0, <2.3",
   "pyro-ppl>=1.9.1",
   "pytest",
   "torch>=2.2.0",
