+++ tests/common.py
@@ -4,9 +4,8 @@
 import numpy as np
 import torch
 
-from cellarium.ml.distributed import GatherLayer
 from cellarium.ml.models import CellariumModel
-from cellarium.ml.utilities.data import get_rank_and_num_replicas
+from cellarium.ml.utilities.distributed import GatherLayer, get_rank_and_num_replicas
 
 
 class BoringDataset(torch.utils.data.Dataset):
